--- telegramBot/cmd/deploy/development/config.yaml
@@ -1,17 +1,0 @@
-total_shutdown_timeout: 5s
-
-grpc_client:
-  host: user-service
-  port: 6000
-
-telegram_bot:
-  token: myToken
-
-logger:
-  file_path: "logs/bot.log"
-  use_local_time: true
-  file_max_size_in_mb: 10
-  file_max_age_in_days: 7
-  level: -1 # trace
-  compress: true
-  file_max_backups: 10
