+++ telegramBot/internal/adapters/grpc/poll_client_impl.go
@@ -1,0 +1,1 @@
+package grpc
