+++ telegramBot/internal/domain/poll.go
@@ -1,0 +1,1 @@
+package domain
