+++ telegramBot/internal/domain/state.go
@@ -1,0 +1,1 @@
+package domain
