+++ telegramBot/internal/domain/user.go
@@ -1,0 +1,1 @@
+package domain
