+++ telegramBot/internal/ports/poll_client.go
@@ -1,0 +1,1 @@
+package ports
