+++ telegramBot/internal/service/bot_service.go
@@ -1,0 +1,1 @@
+package service
