+++ telegramBot/protos/user/proto/user.proto
@@ -1,0 +1,24 @@
+syntax = "proto3";
+
+package auth;
+option go_package = "protos/user/golang";
+
+message RegisterRequest {
+  string phone = 1;
+  string password = 2;
+  string telegramId = 3;
+}
+
+message TelegramLoginRequest {
+  string telegramId = 1;
+}
+
+message CreatorLoginResponse {
+  string id = 1;
+  string token = 2;
+}
+
+service UserService {
+  rpc LoginWithTelegram(TelegramLoginRequest) returns (CreatorLoginResponse);
+  rpc Register(RegisterRequest) returns (CreatorLoginResponse);
+}
